Move pid controller to separate ROS node
