+++ client.launch
@@ -9,4 +9,6 @@
   <param name="carla/port" value="$(arg port)" unless="$(eval port == '')"/>
   <param name="carla/ego_vehicle/role_name" value="$(arg ego_vehicle_name)" unless="$(eval ego_vehicle_name == '')"/>
   <node pkg="carla_ros_bridge" name="carla_ros_bridge" type="client.py" output="screen"/>
+
+  <node pkg="carla_ros_bridge" name="pid_controller" type="pid_controller.py" output="screen"/>
 </launch>
